[ESSENTIAL]Extract CoreData managed model classes into separate files
